+++ packages/client/src/entities/user/hook/useAuth.ts
@@ -1,0 +1,14 @@
+import { makeMapState, useMapState } from '@@shared/lib/model/hooks';
+import { getCurrentUser } from '../model/selectors';
+
+const mapState = makeMapState((state) => ({
+  user: getCurrentUser(state),
+}));
+
+export const useAuth = () => {
+  const {
+    user: { isLoading, isAuth, data },
+  } = useMapState(mapState);
+
+  return { isLoading, isAuth, user: data };
+};
