+++ packages/client/src/entities/user/index.ts
@@ -1,4 +1,5 @@
+export { useAuth } from './hook/useAuth';
 export { authService } from './api/auth.service';
 export { userSlice } from './model/slice';
 export { registration, login, fetchUser } from './model/reducers';
-export { getCurrentUser, useAuth } from './model/selectors';
+export { getCurrentUser } from './model/selectors';
