+++ packages/client/src/shared/lib/HTTPTransport/HTTPTransport.ts
@@ -1,7 +1,7 @@
 import { AxiosInstance } from 'axios';
 import { axiosInstance } from './axios-config';
 
-export abstract class Http {
+export abstract class HTTPTransport {
   protected http: AxiosInstance;
 
   protected constructor(endpoint: string) {
