+++ packages/client/src/shared/lib/HTTPTransport/index.ts
@@ -1,0 +1,1 @@
+export { HTTPTransport } from './HTTPTransport';
