--- packages/client/src/shared/lib/http/index.ts
@@ -1,1 +1,0 @@
-export { Http } from './http';
